ajout de l'info transport
