+++ Bundles/Infos/Screen/Navettes.js
@@ -1,32 +1,21 @@
-
-import React, {Component} from 'react';
-import {Platform, StyleSheet, Text, View} from 'react-native';
-
+import React, { Component } from 'react'
+import { Platform, StyleSheet, Text, View, ScrollView } from 'react-native'
+import Accordion from '../Accordion/Accordion.js'
 
 class Navettes extends React.Component {
   render() {
     return (
       <View style={styles.container}>
-        <Text style={styles.welcome}>TESTTTT</Text>
-        <Text style={styles.instructions}>YOLOOOOOO</Text>
+        <Accordion />
       </View>
-    );
+    )
   }
 }
 
-export default Navettes;
+export default Navettes
 
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: 'red',
   },
-  welcome: {
-    fontSize: 20,
-    textAlign: 'center',
-    margin: 10,
-  },
-
-  });
+})
